Split Monitor module into Internal and main
